pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "carla-driving-toolkit"
version = "0.1.0"
description = "A toolkit for integrating CARLA simulator with driving models"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]

dependencies = [
    "carla>=0.9.13",
    "numpy>=1.21.0",
    "torch>=1.9.0",
    "opencv-python>=4.5.0",
    "matplotlib>=3.4.0",
    "pyyaml>=5.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.0",
    "isort>=5.0",
    "flake8>=3.9",
    "mypy>=0.910",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 